+++ controller/categories/restaurant.js
@@ -2,6 +2,7 @@ const { Restaurant, Offer, validateRestaurant } = require('../../models/categori
 const { User } = require('../../models/user/user');
 const _ = require('lodash');
 const { rest, result } = require('lodash');
+const { Owner } = require('../../models/owner/owner');
 
 exports.getRest = async function (req, res, next) {
 
@@ -50,15 +51,32 @@ exports.postRest = async function (req, res, next) {
     if (error) return res.status(400).send(error.details[0].message);
     let rest = await Restaurant.findOne({ name: req.body.name, lat: req.body.lat, lng: req.body.lng });
     if (rest) return res.status(400).send('this restaurant is already here');
-    rest = new Restaurant(_.pick(req.body,
-        ['name', 'address', 'pic', 'menu', 'workTime','in_Favorites', 'cuisineType', 'city', 'lat', 'lng']));
-    rest = await rest.save();
+    
+       rest = new Restaurant({
+        name:req.body.name,
+        address:req.body.address,
+        pic:req.body.pic,
+        menu:req.body.menu,
+        workTime:req.body.workTime,
+        inWishList:req.body.inWishList,
+        cuisineType:req.body.cuisineType,
+        city:req.body.city,
+        lat:req.body.lat,
+        lng:req.body.lng,
+        owner:req.owner._id
+       })
+    
+     rest = await rest.save();
     res.send(rest);
     next();
 }
 
 exports.putRest = async function (req, res, next) {
-    let rest = await Restaurant.findById(req.params.id);
+
+    let owner = await Owner.findById(req.owner._id)
+    if(!owner)return res.send('you are not owner of item')
+    //let rest = await Restaurant.findById(req.params.id);
+    let rest= await Restaurant.findOne({ _id:req.params.id , owner:req.owner._id});
     if (!rest) return res.status(404).send('Not found check your id');
     const { error } = validateRestaurant(req.body);
     if (error) return res.status(400).send(error.details[0].message);
@@ -75,7 +93,7 @@ exports.putRest = async function (req, res, next) {
             lat: req.body.lat,
             lng: req.body.lng,
             city: req.body.city,
-            in_Favorites: req.body.in_Favorites
+           inWishList: req.body.inWishList
         }
     }, { new: true })
 
@@ -86,10 +104,20 @@ exports.putRest = async function (req, res, next) {
 }
 
 exports.deleteRest = async function (req, res, next) {
-    const rest = await Restaurant.findByIdAndRemove(req.params.id);
-    if (!rest) return res.status(404).send('not found check your id')
+    if(req.user){
+        if(req.user.local.isAdmin){
+
+            const rest = await Restaurant.findByIdAndRemove(req.params.id);
+            if (!rest) return res.status(404).send('not found check your id')
+            res.send(rest);
+        }
+}
+else{
+    const rest = await Restaurant.findOneAndDelete({_id:req.params.id, owner:req.owner.id});
+    if (!rest) return res.status(404).send('not found')
     res.send(rest);
-    next();
+
+}
 }
 
 exports.addComment = async (req, res, next) => {
@@ -121,45 +149,26 @@ exports.addComment = async (req, res, next) => {
 
 
 exports.addReview = async (req, res, next) => {
-    const user = await User.findById(req.body.userId);
     const restaurant = await Restaurant.findById(req.params.id);
     const review = restaurant.review;
-    //console.log(user);
-
-    let userName = function () {
-        let localName = user.local.name;
-        if (localName == null) return user.google.name;
-        else return localName
-    }
-
-
-    let result = review.find((rev) => rev.UserId == req.body.userId)
+    let result = review.find((rev) => JSON.stringify(rev.UserId) == JSON.stringify(req.user._id))
     if (result) {
-        //console.log(result);
         return res.send('cannot send again');
     } else {
         review.push({
-            "name": userName(),
-            "UserId": req.body.userId,
+            "UserId": req.user._id,
             "rate": req.body.rate,
             "comment": req.body.comment,
         })
-
         await restaurant.save();
-
+        
         restaurant.rate = review.reduce((total, num) => {
             return rating = Math.round((total + (num.rate / review.length)) * 10) / 10;
-            //if(rating>5)return rating=5;
-
         }, 0);
-        console.log(restaurant.rate);
-
+        
         await restaurant.save();
 
     }
-
-
-
 
     res.status(200).json({
         "status": true,
@@ -171,39 +180,16 @@ exports.addReview = async (req, res, next) => {
 }
 
 
-        // second true solution
-        /*const findAverage=(arr)=>{
-            return arr.reduce((acc,val)=>{
-                return acc + (val.rate/arr.length);
-
-            },0) 
-        };
-
-        console.log(findAverage(review));
-*/
-
-
-        /*        var arr = [1, 2, 3, 4]
-        function sumOfArr(total, current) {
-            return total + current / arr.length;
-        }
-
-        var sum = arr.reduce(sumOfArr)
-        console.log('sum of arr is', sum);
-*/
-
-        //console.log(rate);
-
 exports.deleteReviewFromRestaurant = async (req, res, next)=>{
     const restaurant = await Restaurant.findById(req.params.id);
-    console.log(restaurant);
+    //console.log(restaurant);
     review = restaurant.review;
-    console.log('review is',review);
-    element = review.find((rev) => rev.UserId == req.body.userId)
-    console.log('element is',element);
+    //console.log('review is',review);
+    element = review.find((rev) => JSON.stringify(rev.UserId) == JSON.stringify(req.user._id));
+    //console.log('element is',element);
    if(!element)return res.status(404).send('User Review Not Found')
     element1 = review.id(req.body.id).remove();
-    console.log('delete review is',element1);
+    //console.log('delete review is',element1);
     await restaurant.save();
 
     res.send('success review deleted');
@@ -212,7 +198,44 @@ exports.deleteReviewFromRestaurant = async (req, res, next)=>{
 }
 
 
-exports.addGetOffer = async (req, res, next) => {
+
+
+exports.booking = async (req, res, next) => {
+    const user = await User.findById(req.user._id);
+    const restaurant = await Restaurant.findById(req.params.id);
+    const booking = restaurant.booking;
+
+    let userName = function () {
+        let localName = user.local.name;
+        if (localName == null) 
+        return user.google.name;
+        else return localName
+    }
+        let result = booking.find((book) => JSON.stringify(book.UserId) ==  JSON.stringify(req.user._id)) //req.body.userId})
+ 
+        if (result) {
+    return res.send('cannot booking again');
+    } else {
+        booking.push({
+            "name": userName(),
+            "UserId": req.user._id,
+            "price": req.body.price,
+            "date": req.body.Date,
+            "payed":true
+        })
+        
+        await restaurant.save();
+    }
+
+    res.status(200).json({
+        "status": true,
+        "message": "booked success",
+        "data": booking
+    })
+}
+
+
+/*exports.addGetOffer = async (req, res, next) => {
     const user = await User.findById(req.params.id);
     //if (!user) return res.send('you must login to get offer');
 
@@ -259,4 +282,4 @@ exports.getRestaurantComments = async (req, res, next) => {
         "message": "success",
         "data": restaurant.comment
     });
-}
+}*/
